requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.3.1
attrs==23.2.0
autopep8==1.5.7
blinker==1.7.0
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.7
Django==3.1.6
django-ckeditor==6.0.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-jazzmin==2.6.0
django-js-asset==1.2.2
django-rest-multiple-models==2.1.3
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-summernote==0.8.11.6
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
filelock==3.13.1
Flask==3.0.0
frozenlist==1.4.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
openapi-codec==1.3.2
packaging==23.1
pandas==2.2.1
Pillow==10.0.1
platformdirs==3.11.0
psycopg2-binary==2.9.6
pycodestyle==2.7.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2021.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.1
sydney-py==0.20.4
threadpoolctl==3.3.0
toml==0.10.2
tqdm==4.66.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.24.6
websockets==12.0
Werkzeug==3.0.1
whitenoise==5.2.0
yarl==1.9.4
